Refactor: jotai atom 모듈화 및 전역 상태 도메인 별 custom hook으로 분할.
