+++ src/entities/post/model/atoms/comments.ts
@@ -1,0 +1,5 @@
+import { atom } from 'jotai';
+
+import type { Comment } from '../types';
+
+export const commentsAtom = atom<Record<number, Comment[]>>({});
