+++ src/entities/post/model/atoms/ui.ts
@@ -1,0 +1,8 @@
+import { atom } from 'jotai';
+
+export const showAddDialogAtom = atom<boolean>(false);
+export const showEditDialogAtom = atom<boolean>(false);
+export const showPostDetailDialogAtom = atom<boolean>(false);
+export const showAddCommentDialogAtom = atom<boolean>(false);
+export const showEditCommentDialogAtom = atom<boolean>(false);
+export const showUserModalAtom = atom<boolean>(false);
