+++ src/entities/post/model/hooks.ts
@@ -1,0 +1,1 @@
+export * from './hooks/index.ts';
