+++ src/entities/post/model/hooks/useDialogState.ts
@@ -1,0 +1,33 @@
+import { useAtom } from 'jotai';
+
+import {
+  showAddDialogAtom,
+  showEditDialogAtom,
+  showAddCommentDialogAtom,
+  showEditCommentDialogAtom,
+  showPostDetailDialogAtom,
+  showUserModalAtom,
+} from '../atoms';
+
+export function useDialogState() {
+  const [showAddDialog, setShowAddDialog] = useAtom(showAddDialogAtom);
+  const [showEditDialog, setShowEditDialog] = useAtom(showEditDialogAtom);
+  const [showAddCommentDialog, setShowAddCommentDialog] = useAtom(showAddCommentDialogAtom);
+  const [showEditCommentDialog, setShowEditCommentDialog] = useAtom(showEditCommentDialogAtom);
+  const [showPostDetailDialog, setShowPostDetailDialog] = useAtom(showPostDetailDialogAtom);
+  const [showUserModal, setShowUserModal] = useAtom(showUserModalAtom);
+  return {
+    showAddDialog,
+    setShowAddDialog,
+    showEditDialog,
+    setShowEditDialog,
+    showAddCommentDialog,
+    setShowAddCommentDialog,
+    showEditCommentDialog,
+    setShowEditCommentDialog,
+    showPostDetailDialog,
+    setShowPostDetailDialog,
+    showUserModal,
+    setShowUserModal,
+  };
+}
