+++ src/entities/post/model/hooks/useSelectionState.ts
@@ -1,0 +1,32 @@
+import { useAtom } from 'jotai';
+
+import {
+  selectedPostAtom,
+  selectedCommentAtom,
+  selectedUserAtom,
+  newPostAtom,
+  newCommentAtom,
+  commentsAtom,
+} from '../atoms';
+
+export function useSelectionState() {
+  const [selectedPost, setSelectedPost] = useAtom(selectedPostAtom);
+  const [selectedComment, setSelectedComment] = useAtom(selectedCommentAtom);
+  const [selectedUser, setSelectedUser] = useAtom(selectedUserAtom);
+  const [newPost, setNewPost] = useAtom(newPostAtom);
+  const [newComment, setNewComment] = useAtom(newCommentAtom);
+  const [comments] = useAtom(commentsAtom);
+  return {
+    selectedPost,
+    setSelectedPost,
+    selectedComment,
+    setSelectedComment,
+    selectedUser,
+    setSelectedUser,
+    newPost,
+    setNewPost,
+    newComment,
+    setNewComment,
+    comments,
+  };
+}
