+++ src/features/PaginationControls/index.tsx
@@ -1,36 +1,21 @@
 import * as SelectPrimitive from '@radix-ui/react-select';
 import React from 'react';
 
+import { usePostsState } from '../../entities/post/model/hooks';
 import { Button, SelectContent, SelectItem, SelectTrigger } from '../../shared/ui';
 
 export const Select = SelectPrimitive.Root;
 export const SelectValue = SelectPrimitive.Value;
 
-export type PaginationControlsProps = {
-  limit: number;
-  skip: number;
-  total: number;
-  onLimitChange: (limit: number) => void;
-  onPrev: () => void;
-  onNext: () => void;
-};
-
-const PaginationControls: React.FC<PaginationControlsProps> = ({
-  limit,
-  skip,
-  total,
-  onLimitChange,
-  onPrev,
-  onNext,
-}) => {
+const PaginationControls: React.FC = () => {
+  const { limit, setLimit, skip, setSkip, total } = usePostsState();
+  const handlePrev = () => setSkip(Math.max(0, skip - limit));
+  const handleNext = () => setSkip(skip + limit);
   return (
     <div className='flex justify-between items-center'>
       <div className='flex items-center gap-2'>
         <span>표시</span>
-        <Select
-          value={limit.toString()}
-          onValueChange={(value: string) => onLimitChange(Number(value))}
-        >
+        <Select value={limit.toString()} onValueChange={(value: string) => setLimit(Number(value))}>
           <SelectTrigger className='w-[180px]'>
             <SelectValue placeholder='10' />
           </SelectTrigger>
@@ -43,10 +28,10 @@ const PaginationControls: React.FC<PaginationControlsProps> = ({
         <span>항목</span>
       </div>
       <div className='flex gap-2'>
-        <Button disabled={skip === 0} onClick={onPrev}>
+        <Button disabled={skip === 0} onClick={handlePrev}>
           이전
         </Button>
-        <Button disabled={skip + limit >= total} onClick={onNext}>
+        <Button disabled={skip + limit >= total} onClick={handleNext}>
           다음
         </Button>
       </div>
