--- src/features/post-manage/index.ts
@@ -1,2 +1,0 @@
-// Unused after cleanup. File intentionally left blank.
-export {};
