+++ src/widgets/CommentAddDialog/index.tsx
@@ -1,38 +1,34 @@
 import * as DialogPrimitive from '@radix-ui/react-dialog';
 import React from 'react';
 
+import {
+  useCommentActions,
+  useDialogState,
+  useSelectionState,
+} from '../../entities/post/model/hooks';
 import { AddCommentFormContent } from '../../features/PostDialogs';
 import { DialogContent, DialogHeader, DialogTitle } from '../../shared/ui';
 
 export const Dialog = DialogPrimitive.Root;
 
-type Props = {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  newComment: { body: string; postId: number | null; userId: number };
-  setNewComment: (v: { body: string; postId: number | null; userId: number }) => void;
-  onSubmit: () => void;
+const CommentAddDialog: React.FC = () => {
+  const { showAddCommentDialog, setShowAddCommentDialog } = useDialogState();
+  const { newComment, setNewComment } = useSelectionState();
+  const { addComment } = useCommentActions();
+  return (
+    <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>새 댓글 추가</DialogTitle>
+        </DialogHeader>
+        <AddCommentFormContent
+          newComment={newComment}
+          setNewComment={setNewComment}
+          onSubmit={addComment}
+        />
+      </DialogContent>
+    </Dialog>
+  );
 };
-
-const CommentAddDialog: React.FC<Props> = ({
-  open,
-  onOpenChange,
-  newComment,
-  setNewComment,
-  onSubmit,
-}) => (
-  <Dialog open={open} onOpenChange={onOpenChange}>
-    <DialogContent>
-      <DialogHeader>
-        <DialogTitle>새 댓글 추가</DialogTitle>
-      </DialogHeader>
-      <AddCommentFormContent
-        newComment={newComment}
-        setNewComment={setNewComment}
-        onSubmit={onSubmit}
-      />
-    </DialogContent>
-  </Dialog>
-);
 
 export default CommentAddDialog;
