+++ src/widgets/PostDetailDialog/index.tsx
@@ -1,37 +1,65 @@
 import * as DialogPrimitive from '@radix-ui/react-dialog';
-import React from 'react';
+import React, { useEffect } from 'react';
 
-import type { Post } from '../../entities/post/model/types';
+import {
+  useCommentActions,
+  useDialogState,
+  usePostsState,
+  useSelectionState,
+} from '../../entities/post/model/hooks';
 import { DialogContent, DialogHeader, DialogTitle } from '../../shared/ui';
+import CommentsList from '../../widgets/CommentsList';
 
 export const Dialog = DialogPrimitive.Root;
 
-type Props = {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  selectedPost: Post | null;
-  searchQuery: string;
-  highlightText: (text: string, highlight: string) => React.ReactNode;
-  renderComments: (postId: number) => React.ReactNode;
-};
+const PostDetailDialog: React.FC = () => {
+  const { showPostDetailDialog, setShowPostDetailDialog } = useDialogState();
+  const { selectedPost, setSelectedComment, newComment, setNewComment, comments } =
+    useSelectionState();
+  const { searchQuery } = usePostsState();
+  const { fetchComments, deleteComment, likeComment } = useCommentActions();
 
-const PostDetailDialog: React.FC<Props> = ({
-  open,
-  onOpenChange,
-  selectedPost,
-  searchQuery,
-  highlightText,
-  renderComments,
-}) => {
+  // 기존 페이지의 하이라이트 로직을 동일하게 사용
+  const highlightText = (text: string | undefined, highlight: string) => {
+    if (!text) return null;
+    if (!highlight.trim()) return <span>{text}</span>;
+    const regex = new RegExp(`(${highlight})`, 'gi');
+    const parts = text.split(regex);
+    return (
+      <span>
+        {parts.map((part, i) =>
+          regex.test(part) ? <mark key={i}>{part}</mark> : <span key={i}>{part}</span>,
+        )}
+      </span>
+    );
+  };
+
+  useEffect(() => {
+    if (showPostDetailDialog && selectedPost) {
+      fetchComments(selectedPost.id);
+    }
+  }, [showPostDetailDialog, selectedPost]);
+
   return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
+    <Dialog open={showPostDetailDialog} onOpenChange={setShowPostDetailDialog}>
       <DialogContent className='max-w-3xl'>
         <DialogHeader>
           <DialogTitle>{highlightText(selectedPost?.title ?? '', searchQuery)}</DialogTitle>
         </DialogHeader>
         <div className='space-y-4'>
           <p>{highlightText(selectedPost?.body ?? '', searchQuery)}</p>
-          {selectedPost && renderComments(selectedPost.id)}
+          {selectedPost && (
+            <CommentsList
+              postId={selectedPost.id}
+              comments={comments}
+              searchQuery={searchQuery}
+              onAddClick={(pid) => setNewComment({ ...newComment, postId: pid })}
+              onLike={likeComment}
+              onEdit={(comment) => setSelectedComment(comment)}
+              onDelete={deleteComment}
+              highlightText={(text, h) => highlightText(text, h)}
+            />
+          )}
         </div>
       </DialogContent>
     </Dialog>
