+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/index.tsx","./src/main.tsx","./src/apps/app.tsx","./src/entities/post/api/index.ts","./src/entities/post/model/atoms.ts","./src/entities/post/model/types.ts","./src/features/paginationcontrols/index.tsx","./src/features/postdialogs/index.tsx","./src/features/postscontrols/index.tsx","./src/features/post-manage/index.ts","./src/pages/postsmanagerpage.tsx","./src/shared/ui/button.tsx","./src/shared/ui/card.tsx","./src/shared/ui/cardcontent.tsx","./src/shared/ui/cardheader.tsx","./src/shared/ui/cardtitle.tsx","./src/shared/ui/dialogcontent.tsx","./src/shared/ui/dialogheader.tsx","./src/shared/ui/dialogtitle.tsx","./src/shared/ui/input.tsx","./src/shared/ui/selectcontent.tsx","./src/shared/ui/selectitem.tsx","./src/shared/ui/selecttrigger.tsx","./src/shared/ui/table.tsx","./src/shared/ui/tablebody.tsx","./src/shared/ui/tablecell.tsx","./src/shared/ui/tablehead.tsx","./src/shared/ui/tableheader.tsx","./src/shared/ui/tablerow.tsx","./src/shared/ui/textarea.tsx","./src/shared/ui/index.ts","./src/types/global.d.ts","./src/widgets/commentslist/index.tsx","./src/widgets/footer/index.ts","./src/widgets/header/index.ts","./src/widgets/postdetaildialog/index.tsx","./src/widgets/poststable/index.tsx","./src/widgets/usermodal/index.tsx"],"version":"5.9.2"}
+{"root":["./src/app.tsx","./src/index.tsx","./src/main.tsx","./src/apps/app.tsx","./src/entities/post/api/index.ts","./src/entities/post/model/atoms.ts","./src/entities/post/model/hooks.ts","./src/entities/post/model/types.ts","./src/entities/post/model/atoms/actions.ts","./src/entities/post/model/atoms/comments.ts","./src/entities/post/model/atoms/drafts.ts","./src/entities/post/model/atoms/index.ts","./src/entities/post/model/atoms/selection.ts","./src/entities/post/model/atoms/state.ts","./src/entities/post/model/atoms/ui.ts","./src/entities/post/model/hooks/index.ts","./src/entities/post/model/hooks/usecommentactions.ts","./src/entities/post/model/hooks/usedialogstate.ts","./src/entities/post/model/hooks/usepostactions.ts","./src/entities/post/model/hooks/usepostsstate.ts","./src/entities/post/model/hooks/useselectionstate.ts","./src/features/paginationcontrols/index.tsx","./src/features/postdialogs/index.tsx","./src/features/postscontrols/index.tsx","./src/pages/postsmanagerpage.tsx","./src/shared/ui/button.tsx","./src/shared/ui/card.tsx","./src/shared/ui/cardcontent.tsx","./src/shared/ui/cardheader.tsx","./src/shared/ui/cardtitle.tsx","./src/shared/ui/dialogcontent.tsx","./src/shared/ui/dialogheader.tsx","./src/shared/ui/dialogtitle.tsx","./src/shared/ui/input.tsx","./src/shared/ui/selectcontent.tsx","./src/shared/ui/selectitem.tsx","./src/shared/ui/selecttrigger.tsx","./src/shared/ui/table.tsx","./src/shared/ui/tablebody.tsx","./src/shared/ui/tablecell.tsx","./src/shared/ui/tablehead.tsx","./src/shared/ui/tableheader.tsx","./src/shared/ui/tablerow.tsx","./src/shared/ui/textarea.tsx","./src/shared/ui/index.ts","./src/types/global.d.ts","./src/widgets/commentadddialog/index.tsx","./src/widgets/commenteditdialog/index.tsx","./src/widgets/commentslist/index.tsx","./src/widgets/footer/index.tsx","./src/widgets/header/index.tsx","./src/widgets/postadddialog/index.tsx","./src/widgets/postdetaildialog/index.tsx","./src/widgets/posteditdialog/index.tsx","./src/widgets/postsheader/index.tsx","./src/widgets/poststable/index.tsx","./src/widgets/usermodal/index.tsx"],"version":"5.9.2"}
